Feature/host

+++ onelove/api/service.py
@@ -7,23 +7,17 @@ from ..models.auth import User
 from ..models.service import Service
 from .mixins import ServiceMixin
 from .namespaces import ns_service
+from .pagination import paginate, parser
 from .resources import ProtectedResource
 from .schemas import ServiceSchema
 
 
 @ns_service.route('', endpoint='services')
 class ServiceListAPI(ProtectedResource):
+    @ns_service.expect(parser)
     def get(self):
         """List services"""
-        services = Service.objects.all()
-
-        schema = ServiceSchema(many=True)
-        response, errors = schema.dump(services)
-
-        if errors:
-            abort(409, errors)
-
-        return response, 200
+        return paginate(Service.objects(), ServiceSchema())
 
     @ns_service.response(409, 'Service with that name already exists')
     @ns_service.expect(ServiceSchema.fields())
@@ -32,48 +26,53 @@ class ServiceListAPI(ProtectedResource):
         email = get_jwt_identity()
         user = User.objects.get(email=email)
         schema = ServiceSchema()
-        data, errors = schema.load(current_app.api.payload)
+        service, errors = schema.load(current_app.api.payload)
         if errors:
             return errors, 409
-        service = Service(name=data.name, user=user.pk)
+        service.user = user.pk
         try:
             service.save()
         except NotUniqueError:
             abort(409, error='Service with that name already exists')
-        response = schema.dump(service)
-
-        return response, 201
+        response, errors = schema.dump(service)
+        if errors:
+            return errors, 409
+        return response
 
 
 @ns_service.route('/<id>', endpoint='services_service')
 @ns_service.response(404, 'Service not found')
 class ServiceAPI(ProtectedResource, ServiceMixin):
-    @ns_service.marshal_with(ServiceSchema.fields())
     def get(self, id):
         """Show service details"""
-        service = self._find_service(id)
-        return service
+        service = self.find_service(id)
+        schema = ServiceSchema()
+        response, errors = schema.dump(service)
+        if errors:
+            return errors, 409
+        return response
 
     @ns_service.expect(ServiceSchema.fields())
     def patch(self, id):
         """Update service"""
-        service = self._find_service(id)
+        service = self.find_service(id)
         schema = ServiceSchema()
         data, errors = schema.load(current_app.api.payload)
         if errors:
             return errors, 409
         service.name = data.name or service.name
         service.save()
-
-        response = schema.dump(service)
-
+        response, errors = schema.dump(service)
+        if errors:
+            return errors, 409
         return response
 
-    @ns_service.marshal_with(ServiceSchema.fields())
     def delete(self, id):
         """Delete the service."""
-        service = self._find_service(id)
-
+        service = self.find_service(id)
+        schema = ServiceSchema()
+        response, errors = schema.dump(service)
+        if errors:
+            return errors, 409
         service.delete()
-
-        return service
+        return response
